Added range removal functionality to GuestMemory
Added a function called `remove_range` to the `GuestMemory` trait. This
should tell the host kernel that it is safe to reclaim some range of
memory. This would be useful for implementing a memory balloon device.

Also, fixed a typo in the `guest_memory.rs` comments, and a typo in
`DESIGN.md`.

Signed-off-by: Tamio-Vesa Nakajima <tamiove@amazon.com>